+++ API-Restful/API-Restful/Business/Interfaces/IPersonService.cs
@@ -6,12 +6,12 @@ namespace API_Restful.Business.Interfaces
 	{
 		Person Create(Person person);
 
-		Person FindById(long id);
+		Person FindById(int id);
 
 		List<Person> FindAll();
 
 		Person Update(Person person);
 
-		void Delete(long id);
+		void Delete(int id);
 	}
 }
